Move FoldingRange to its own file
No code change here.
